Moved functions from index.jsx to form to make them work in routes
